% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{angle}
\alias{angle}
\alias{degree}
\alias{radian}
\alias{turn}
\title{angle class}
\usage{
angle(.data = class_missing)

degree(degree = class_missing)

radian(radian = class_missing)

turn(turn = class_missing)
}
\arguments{
\item{degree}{degrees}

\item{radian}{radians}

\item{turn}{proportion of full turns of a circle (1 turn = 2 * pi radians)}

\item{gradian}{gradians, gons, or grads (right angle = 100 gradians)}
}
\description{
Creates an angle in the metric of radians, degrees, turns, and gradians.
}
\details{
Angles turns can be any real number, but degrees are displayed as values between -360 and +360,
radians are between -2pi and +2pi, and gradians are between -400 and +400.
}
\examples{
# Trhee Different ways to make a right angle
## A quarter turn
turn(.25)

## half pi radians
radian(.5 * pi)

## 90 degrees
degree(90)

# Operations
angle(degree = 30) + angle(degree = 20)
angle(degree = 350) + angle(degree = 20)
angle(degree = 30) - angle(degree = 20)
angle(degree = 30) - angle(degree = 50)

2 * angle(degree = 30)
angle(degree = 30) / 3

angle(degree = 30) + pi # added or subtracted numbers are radians
radian(1) + 1 # added or subtracted numbers are radians
degree(30) + 180 # added or subtracted numbers are degrees
turn(.25) + .25 # added or subtracted numbers are turns
gradian(50) + 50 # added or subtracted numbers are turns

Trigonometric functions work as normal
x <- angle(degree = 180)
sin(x)
cos(x)
tan(x)
}
