% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{as.geom}
\alias{as.geom}
\title{Convert shapes to ggplot2 geoms}
\usage{
as.geom(x, ...)
}
\arguments{
\item{x}{a shape}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Pass arguments to ggplot2::geom_point}
}
\description{
Convert shapes to ggplot2 geoms
}
