% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{bind}
\alias{bind}
\title{bind method}
\usage{
bind(x, ...)
}
\arguments{
\item{x}{list of objects to bind}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\description{
bind method
}
\examples{
bind(c(point(1,2), point(3,4)))
bind(c(circle(point(0,0), radius = 1),
       circle(point(1,1), radius = 2)))
}
