% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{bind_shape}
\alias{bind_shape}
\title{Combine several objects (of the same type) into one object}
\usage{
bind_shape(x, ...)
}
\arguments{
\item{x}{object}
}
\description{
Combine several objects (of the same type) into one object
}
