% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzcurve.R
\name{bzcurve}
\alias{bzcurve}
\title{The bzcurve (i.e., bezier curve) class}
\usage{
bzcurve(
  p = class_missing,
  label = class_missing,
  n = 360,
  alpha = class_missing,
  arrow_head = class_missing,
  arrow_fins = class_missing,
  arrowhead_length = class_missing,
  length_head = class_missing,
  length_fins = class_missing,
  color = class_missing,
  fill = class_missing,
  lineend = class_missing,
  linejoin = class_missing,
  linewidth = 0.75,
  linewidth_fins = class_missing,
  linewidth_head = class_missing,
  linetype = class_missing,
  resect = class_missing,
  resect_fins = class_missing,
  resect_head = class_missing,
  stroke_color = class_missing,
  stroke_width = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{p}{point object or list of point objects}

\item{style}{Gets and sets the styles associated with bzcurves}

\item{...}{properties passed to style}

\item{length}{The number of curves in the bzcurve object}

\item{tibble}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggarrow::geom_arrow}.}
}
\description{
The bzcurve is specified with a point object that contains at least 2 points, the start and the end. Such a "curve" would actually be a straight line segment. If three points are specified, the middle point is a control point, and a quadratic bezier curve will result. Higher-order bezier curves can be created by having more control points in the middle.
}
\details{
If you wish to specify multiple bezier curves, you must supply a list of point objects. When plotted, the bzcurve function uses the bezier::bezier function to create the point coordinates of the curve and the ggarrow::geom_arrow function to create the geom.
}
