% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcs.R
\name{circle_pie}
\alias{circle_pie}
\title{circle_pie}
\usage{
circle_pie(
  center = ob_point(0, 0),
  radius = 1,
  proportion_list = list(c(0.5, 0.5)),
  start = degree(0)
)
}
\arguments{
\item{center}{center point of pie}

\item{radius}{length of circle radius}

\item{proportion_list}{list of proportions}

\item{start}{angle at which pie wedges start}

\item{color_list}{color of wedge borders}

\item{fill_list}{color of wedge fills}
}
\description{
pie wedges
}
