% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{class_aesthetics_list}
\alias{class_aesthetics_list}
\title{class_aesthetics_list}
\usage{
class_aesthetics_list(
  geom = class_missing,
  style = class_missing,
  mappable_bare = class_missing,
  mappable_identity = class_missing,
  not_mappable = class_missing,
  required_aes = class_missing,
  omit_names = class_missing,
  inherit.aes = class_missing
)
}
\arguments{
\item{geom}{Which geom function converts the shape}

\item{style}{vector of style names}

\item{mappable_bare}{aesthetics used without identity function}

\item{mappable_identity}{aesthetics used with identity function}

\item{not_mappable}{properties that cannot be mapped and thus are created with separate geom objects for each unique combination of values}

\item{required_aes}{required aesthetics}

\item{omit_names}{properties that are ignored}

\item{inherit.aes}{Defaults to \code{FALSE} so that ggdiagram objects do not interfere with other layers in the ggplot}
}
\description{
list of aesthetics
}
\keyword{internal}
