% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{class_color}
\alias{class_color}
\title{color class}
\usage{
class_color(color = class_missing)
}
\arguments{
\item{color}{character (R color or hex code)}

\item{transparentize}{function to return the color with a new transparency (i.e., alpha)}

\item{lighten}{function to return a lighter color}

\item{darken}{function to return a darker color}
}
\description{
color class
}
\examples{
mycolor <- class_color("blue")
mycolor
c(mycolor)
mycolor@tranparentize(.5)
mycolor@lighten(.5)
mycolor@darken(.5)
}
