% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{class_margin}
\alias{class_margin}
\title{Create a structure for label margins and label padding}
\usage{
class_margin(x = numeric(0), units = "pt")
}
\arguments{
\item{x}{A vector of numeric data, grid}
}
\description{
Create a structure for label margins and label padding
}
\keyword{internal}
