% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{connect}
\alias{connect}
\title{Arrow connect one shape to another}
\usage{
connect(x, y, ...)
}
\arguments{
\item{x}{first shape (e.g., point, circle, ellipse, rectangle)}

\item{y}{second shape}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Arguments passed to style}
}
\description{
Arrow connect one shape to another
}
