% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{get_tibble}
\alias{get_tibble}
\alias{get_tibble_defaults}
\title{Get object data with styles in a tibble}
\usage{
get_tibble(x)

get_tibble_defaults(x)
}
\arguments{
\item{x}{object}
}
\value{
a \link[tibble:tibble]{tibble::tibble}

a \link[tibble:tibble]{tibble::tibble}
}
\description{
Get object data with styles in a tibble

Get object data in a tibble, filling in any missing styles with defaults
}
