% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{gg_arrow}
\alias{gg_arrow}
\title{Create straight arrow connecting two shapes}
\usage{
gg_arrow(shapes = NULL, connection1 = "near", connection2 = "near",
  size = 1, color = "black", type = "open", shiftx = 0, shifty = 0,
  shiftxend = 0, shiftyend = 0, dodge = c(0, 0), linetype = 1)
}
\arguments{
\item{shapes}{Numerical vector of length 2 specifying shapes to connect}

\item{connection1}{Connection on shape 1. Can be "near" for nearest point
to other shape or "up", "down", "left", "right" for top, bottom, left, or 
right side of shape. Can also be a number from 1 to 360 specifying specific
point along the shape, where 1 corresponds to the top.}

\item{connection2}{See above.}

\item{size}{Thickness of outline.}

\item{color}{Color of outline.}

\item{type}{Arrow head type. "open" or "closed".}

\item{shiftx}{distance to shift x coordinate of arrow start}

\item{shifty}{distance to shift y coordinate of arrow start}

\item{shiftxend}{distance to shift x coordinate of arrow end}

\item{shiftyend}{distance to shift y coordinate of arrow end}

\item{dodge}{Numeric vector of length 2 giving degrees to adjust 
arrow start and arrow end.}

\item{linetype}{Line type.}
}
\description{
Create straight arrow connecting two shapes
}
