% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{gg_circle}
\alias{gg_circle}
\title{Create a circle}
\usage{
gg_circle(center = c(0, 0), radius = 10, size = 0.5, color = "black",
  fill = "white", alpha = 1, linetype = 1, label = "",
  label_color = "black", label_size = 5, rotate = 0,
  rotateabout = center)
}
\arguments{
\item{center}{Numerical vector of length 2 specifying x, y coordinates of circle center}

\item{radius}{Radius of circle.}

\item{size}{Thickness of outline.}

\item{color}{Color of outline.}

\item{fill}{Fill color of shape.}

\item{alpha}{Transparency of fill.}

\item{linetype}{Line type.}

\item{label}{Label at center of shape.}

\item{label_color}{Text color of label.}

\item{label_size}{Text size of label.}

\item{rotate}{Degrees to rotate shape.}

\item{rotate_about}{Coordinates to rotate about.}
}
\description{
Create a circle
}
