% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{gg_curve}
\alias{gg_curve}
\title{Create curved arrow connecting two shapes}
\usage{
gg_curve(shapes = NULL, connection1 = 1, connection2 = 1, size = 1,
  color = "black", type = "open", angle = 0, curvature = 0, ncp = 10,
  shiftx = 0, shifty = 0, shiftxend = 0, shiftyend = 0, dodge = c(0,
  0), linetype = 1)
}
\arguments{
\item{shapes}{Numerical vector of length 2 specifying shapes to connect}

\item{connection1}{Connection on shape 1. Can be "near" for nearest point
to other shape or "up", "down", "left", "right" for top, bottom, left, or 
right side of shape. Can also be a number from 1 to 360 specifying specific
point along the shape, where 1 corresponds to the top.}

\item{connection2}{See above.}

\item{size}{Thickness of outline.}

\item{color}{Color of outline.}

\item{type}{Arrow head type. "open" or "closed".}

\item{angle}{A numeric value between 0 and 180, giving an amount 
to skew the control points of the curve. Values less than 90 skew 
the curve towards the start point and values greater than 90 skew 
the curve towards the end point.}

\item{curvature}{A numeric value giving the amount of curvature. 
Negative values produce left-hand curves, positive values produce 
right-hand curves, and zero produces a straight line.}

\item{ncp}{The number of control points used to draw the curve. 
More control points creates a smoother curve.}

\item{shiftx}{distance to shift x coordinate of arrow start}

\item{shifty}{distance to shift y coordinate of arrow start}

\item{shiftxend}{distance to shift x coordinate of arrow end}

\item{shiftyend}{distance to shift y coordinate of arrow end}

\item{dodge}{Numeric vector of length 2 giving degrees to adjust 
arrow start and arrow end.}

\item{linetype}{Line type.}
}
\description{
Create curved arrow connecting two shapes
}
