% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{gg_ellipse}
\alias{gg_ellipse}
\title{Create an ellipse}
\usage{
gg_ellipse(center = c(0, 0), radiusx = 10, radiusy = 5, size = 0.5,
  color = "black", fill = "white", alpha = 1, linetype = 1,
  label = "", label_color = "black", label_size = 5, rotate = 0,
  rotateabout = center)
}
\arguments{
\item{center}{Numerical vector of length 2 specifying x, y coordinates of ellipse center}

\item{radiusx}{Horizontal radius of ellipse.}

\item{radiusy}{Vertical radius of ellipse.}

\item{size}{Thickness of outline.}

\item{color}{Color of outline.}

\item{fill}{Fill color of shape.}

\item{alpha}{Transparency of fill.}

\item{linetype}{Line type.}

\item{label}{Label at center of shape.}

\item{label_color}{Text color of label.}

\item{label_size}{Text size of label.}

\item{rotate}{Degrees to rotate shape.}

\item{rotate_about}{Coordinates to rotate about.}
}
\description{
Create an ellipse
}
