% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{gg_point}
\alias{gg_point}
\title{Create an empty point in plot in which arrows can connect 
to/from (i.e. for creating bend in arrow)}
\usage{
gg_point(center = c(0, 0), label = "", label_color = "black",
  label_size = 5, rotate = 0, rotateabout = center)
}
\arguments{
\item{center}{Numerical vector of length 2 specifying x, y coordinates of point.}

\item{label}{Label at center of shape.}

\item{label_color}{Text color of label.}

\item{label_size}{Text size of label.}

\item{rotate}{Degrees to rotate shape.}

\item{rotate_about}{Coordinates to rotate about.}
}
\description{
Create an empty point in plot in which arrows can connect 
to/from (i.e. for creating bend in arrow)
}
