% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{gg_rect}
\alias{gg_rect}
\title{Create an ellipse}
\usage{
gg_rect(center = c(0, 0), height = 10, width = 5, label = "",
  color = "black", size = 0.5, fill = "white", alpha = 1,
  linetype = 1, label_color = "black", label_size = 5, rotate = 0,
  rotateabout = center)
}
\arguments{
\item{center}{Numerical vector of length 2 specifying x, y coordinates of ellipse center}

\item{height}{Height of rectangle.}

\item{width}{Width of rectangle.}

\item{label}{Label at center of shape.}

\item{color}{Color of outline.}

\item{size}{Thickness of outline.}

\item{fill}{Fill color of shape.}

\item{alpha}{Transparency of fill.}

\item{linetype}{Line type.}

\item{label_color}{Text color of label.}

\item{label_size}{Text size of label.}

\item{rotate}{Degrees to rotate shape.}

\item{rotate_about}{Coordinates to rotate about.}
}
\description{
Create an ellipse
}
