% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram-package.R
\docType{package}
\name{ggdiagram-package}
\alias{ggdiagram}
\alias{ggdiagram-package}
\title{ggdiagram: Object-oriented diagram plots with ggplot2}
\description{
Makes path diagrams via S7 and ggplot2
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjschne/ggdiagram}
  \item \url{https://wjschne.github.io/ggdiagram/}
  \item Report bugs at \url{https://github.com/wjschne/ggdiagram/issues}
}

}
\author{
\strong{Maintainer}: W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})

}
\keyword{internal}
