% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{ggdiagram}
\alias{ggdiagram}
\title{ggdiagram function}
\usage{
ggdiagram(
  font_family = "sans",
  font_size = 11,
  linewidth = 0.5,
  rect_linewidth = linewidth,
  theme_function = ggplot2::theme_void,
  ...
)
}
\arguments{
\item{font_family}{font family}

\item{font_size}{font size in points}

\item{linewidth}{line width}

\item{rect_linewidth}{line with of rectangles}

\item{theme_function}{ggplot2 theme}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Arguments sent to ggplot2::theme}
}
\description{
This is a convenient way to specify geom defaults
}
\examples{
ggdiagram() + ob_circle()
}
