% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside.R
\name{inside}
\alias{inside}
\title{is an ob_point inside a shape ?}
\usage{
inside(x, y)
}
\arguments{
\item{x}{object}

\item{y}{object}
}
\value{
numeric vector where 1 = inside, 0 = on, -1 = outside
}
\description{
is an ob_point inside a shape ?
}
