% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersection}
\alias{intersection}
\title{intersection of 2 objects (e.g., lines)}
\usage{
intersection(x, y, ...)
}
\arguments{
\item{x}{object}

\item{y}{object}
}
\description{
intersection of 2 objects (e.g., lines)
}
