% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label}
\alias{label}
\title{label class}
\usage{
label(
  label = class_missing,
  p = class_missing,
  angle = class_missing,
  alpha = class_missing,
  color = class_missing,
  family = class_missing,
  fill = class_missing,
  fontface = class_missing,
  hjust = class_missing,
  label.color = class_missing,
  label.margin = ggplot2::margin(1, 1, 1, 1, "pt"),
  label.padding = ggplot2::margin(2, 2, 2, 2, "pt"),
  label.r = class_missing,
  label.size = class_missing,
  lineheight = class_missing,
  polar_just = class_missing,
  nudge_x = class_missing,
  nudge_y = class_missing,
  size = class_missing,
  text.color = class_missing,
  vjust = class_missing,
  style = class_missing,
  plot_point = FALSE,
  position = 0.5,
  ...
)
}
\arguments{
\item{label}{text label}

\item{p}{point}

\item{style}{a style list}

\item{...}{properties passed to style}
}
\description{
label class
}
