% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{label_object}
\alias{label_object}
\title{Automatic label for objects}
\usage{
label_object(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> additional arguments}
}
\value{
string
}
\description{
Automatic label for objects
}
