% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{latex_color}
\alias{latex_color}
\title{Surround TeX expression with a color command}
\usage{
latex_color(x, color)
}
\arguments{
\item{x}{TeX expression}

\item{color}{color}
}
\value{
string
}
\description{
Surround TeX expression with a color command
}
\examples{
latex_color("X^2", "red")
}
