% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{lead_cycle}
\alias{lead_cycle}
\alias{lag_cycle}
\title{Finds the "previous" (lag) or "next" (lead) values in a vector or object with values at the end of the vector recycled to the beginning.}
\usage{
lead_cycle(x, n = 1L)

lag_cycle(x, n = 1L)
}
\arguments{
\item{x}{A vector or ggdiagram object}

\item{n}{Positive integer of length 1, giving the number of positions to lag or lead by}
}
\value{
A vector with the same type and size as x but with elements shifted and cycled by n.
}
\description{
Finds the "previous" (lag) or "next" (lead) values in a vector or object with values at the end of the vector recycled to the beginning.
}
\examples{
lead_cycle(1:5)
lead_cycle(1:5, 2)
lag_cycle(1:5)
lag_cycle(1:5, 2)
octagon <- ob_ngon(n = 8)
vertices <- octagon@vertices
ggdiagram() +
  vertices +
  connect(vertices, lead_cycle(vertices),  resect = 2) +
  connect(vertices, lag_cycle(vertices, n = 2),  resect = 2) +
  connect(vertices, lead_cycle(vertices, n = 3),  resect = 2)
}
