% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{map2_ob}
\alias{map2_ob}
\title{Map over two ggdiagram objects}
\usage{
map2_ob(.x, .y, .f, ..., .progress = FALSE)
}
\arguments{
\item{.x}{a ggdiagram object}

\item{.y}{a ggdiagram object}

\item{.f}{a function that returns a ggdiagram object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to .f}

\item{.progress}{display progress if TRUE}
}
\value{
a ggdiagram object
}
\description{
A wrapper for \link[purrr:map2]{purrr::map2}. It takes two ggdiagram objects with multiple elements, applies a function to each element within the objects, and returns a ggdiagram object
}
