% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{mean_color}
\alias{mean_color}
\title{Average across colors}
\usage{
mean_color(x)
}
\arguments{
\item{x}{color}
}
\value{
string
}
\description{
Average across colors
}
\examples{
color_A <- "dodgerblue"
color_B <- "violet"
mean_color(c(color_A, color_B))
}
