% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{midpoint}
\alias{midpoint}
\title{Get one or more points at positions from 0 to 1}
\usage{
midpoint(x, y, position = 0.5, ...)
}
\arguments{
\item{x}{object}

\item{y}{object (can be omitted for segments and arcs)}

\item{position}{numeric vector. 0 is start, 1 is end. Defaults to .5}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\value{
ob_point
}
\description{
It is possible to get more than one midpoint by specifying a position vector with a length greater than 1. Position values outside 0 and 1 will usually work, but will be outside the object.
}
