% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{nudge}
\alias{nudge}
\title{Make an automatic label}
\usage{
nudge(object, x, y, ...)
}
\arguments{
\item{object}{object}

\item{x}{nudge right and left}

\item{y}{nudge up and down}
}
\description{
Make an automatic label
}
