% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{ob_array}
\alias{ob_array}
\title{Object Arrays}
\usage{
ob_array(x, k = 2, sep = 1, where = "east", anchor = "center", ...)
}
\arguments{
\item{x}{shape}

\item{k}{number of duplicate shapes to make}

\item{sep}{separation distance between shapes}

\item{where}{angle or named direction (e.g.,northwest, east, below, left)}

\item{anchor}{bounding box anchor}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to shape}
}
\value{
An array of shapes of the same class as object passed to x
}
\description{
Make an array of shapes along a line
}
\section{Methods}{
\code{ob_array} is an S7 generic with methods available for the following classes:

\itemize{
\item \code{ggdiagram::ob_circle}
\item \code{ggdiagram::ob_ellipse}
\item \code{ggdiagram::ob_point}
\item \code{ggdiagram::ob_rectangle}
}
}

