% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{ob_intercept}
\alias{ob_intercept}
\title{ob_intercept}
\usage{
ob_intercept(
  center = class_missing,
  width = 1,
  label = class_missing,
  top = class_missing,
  left = class_missing,
  right = class_missing,
  radius = class_missing,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{center}{point at center}

\item{width}{length of side}

\item{label}{A character, angle, or label object}

\item{top}{Top vertex of triangle}

\item{left}{Left vertex of triangle}

\item{right}{Right vertex of triangle}

\item{radius}{A numeric or unit vector of length one, specifying the vertex radius}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{Gets and sets the styles associated with polygons}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\description{
Triangle polygons used in path diagrams.
}
\section{Slots}{

\describe{
\item{\code{length}}{The number of polygons in the ob_polygon object}

\item{\code{tibble}}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggplot2::geom_polygon}.}
}}

