% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{ob_label}
\alias{ob_label}
\title{ob_label class}
\usage{
ob_label(
  label = class_missing,
  p = class_missing,
  angle = class_missing,
  alpha = class_missing,
  color = class_missing,
  family = class_missing,
  fill = class_missing,
  fontface = class_missing,
  hjust = class_missing,
  label.color = class_missing,
  label.margin = ggplot2::margin(1, 1, 1, 1, "pt"),
  label.padding = ggplot2::margin(2, 2, 2, 2, "pt"),
  label.r = class_missing,
  label.size = class_missing,
  lineheight = class_missing,
  polar_just = class_missing,
  nudge_x = class_missing,
  nudge_y = class_missing,
  size = class_missing,
  straight = class_missing,
  text.color = class_missing,
  vjust = class_missing,
  style = class_missing,
  plot_point = FALSE,
  position = 0.5,
  x = class_missing,
  y = class_missing,
  ...
)
}
\arguments{
\item{label}{text label}

\item{p}{ob_point}

\item{angle}{angle of text}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{family}{font family}

\item{fill}{character string for fill color}

\item{fontface}{Can be plain, bold, italic, or bold.italic}

\item{hjust}{horizontal justification.}

\item{label.color}{Color of label outline.}

\item{label.margin}{Amount of distance around label. Unit vector of length four. Usually created with \code{ggplot2::margin}.}

\item{label.padding}{Amount of padding around label. Unit vector of length four. Usually created with \code{ggplot2::margin}.}

\item{label.r}{Radius of rounded corners. Defaults to 0.15 lines.}

\item{label.size}{Width of label outline.}

\item{lineheight}{Height of line of text}

\item{polar_just}{an angle, polar point, or point that alters hjust and vjust (polar polar_just not stored in style)}

\item{nudge_x}{Horizontal adjustment to nudge labels by.}

\item{nudge_y}{Vertical adjustment to nudge labels by.}

\item{size}{numeric size}

\item{straight}{logical. If TRUE, make bzpath label text straight instead of curved.}

\item{text.color}{Color of label text.}

\item{vjust}{vertical justification}

\item{style}{a style list}

\item{plot_point}{plot ob_point (default = FALSE)}

\item{position}{position (used in conjunction with the \code{place} function)}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\description{
ob_label class
}
