% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{ob_latex}
\alias{ob_latex}
\title{ob_latex class}
\usage{
ob_latex(
  tex = class_missing,
  center = class_missing,
  width = class_missing,
  height = class_missing,
  border = class_missing,
  family = class_missing,
  filename = class_missing,
  color = class_missing,
  force_recompile = class_missing
)
}
\arguments{
\item{tex}{LaTeX equation}

\item{center}{ob_point}

\item{width}{width (specify width or height but not both)}

\item{height}{height (specify width or height but not both)}

\item{border}{border space (in points) around image}

\item{family}{font family (installed on system) of plain text}

\item{filename}{bare file name without extension (e.g., \code{myequation})}

\item{force_recompile}{Will re-run xelatex even if .pdf file exists already}
}
\description{
make a latex equation
}
