% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{ob_line}
\alias{ob_line}
\title{ob_line class}
\usage{
ob_line(
  slope = numeric(0),
  intercept = numeric(0),
  xintercept = numeric(0),
  a = numeric(0),
  b = numeric(0),
  c = numeric(0),
  alpha = numeric(0),
  color = character(0),
  lineend = numeric(0),
  linejoin = numeric(0),
  linewidth = numeric(0),
  linetype = numeric(0),
  style = S7::class_missing,
  id = character(0),
  ...
)
}
\arguments{
\item{slope}{coefficient in y = slope * x + intercept}

\item{intercept}{value of y when x is 0}

\item{xintercept}{value of x when y is 0}

\item{a}{coefficient in general form: a * x + b * y + c = 0}

\item{b}{coefficient in general form: a * x + b * y + c = 0}

\item{c}{constant in general form: a * x + b * y + c = 0}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{an \link{ob_style} object}

\item{id}{character string to identify object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\value{
ob_line object
}
\description{
Creates a line
}
