% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{ob_polygon}
\alias{ob_polygon}
\title{The ob_polygon (polygon) class}
\usage{
ob_polygon(
  p = class_missing,
  label = class_missing,
  radius = class_missing,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{p}{ob_point or list of ob_point objects}

\item{label}{A character, angle, or label object}

\item{radius}{A numeric or unit vector of length one,  specifying the corner radius}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{Gets and sets the styles associated with polygons}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\description{
A polygon is specified with an obpoint that contains at least 3 points, the start and the end. Any number of intermediate points are possible.
}
\details{
If you wish to specify multiple polygons, you must supply a list of ob_points. When plotted, the ob_polygon function uses the ggplot2::geom_polygon function to create the geom.
}
\section{Slots}{

\describe{
\item{\code{length}}{The number of polygons in the ob_polygon object}

\item{\code{tibble}}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggplot2::geom_polygon}.}
}}

