% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{ob_rectangle}
\alias{ob_rectangle}
\title{ob_rectangle class}
\usage{
ob_rectangle(
  center = S7::class_missing,
  width = numeric(0),
  height = numeric(0),
  east = S7::class_missing,
  north = S7::class_missing,
  west = S7::class_missing,
  south = S7::class_missing,
  northeast = S7::class_missing,
  northwest = S7::class_missing,
  southwest = S7::class_missing,
  southeast = S7::class_missing,
  angle = numeric(0),
  vertex_radius = numeric(0),
  label = character(0),
  alpha = numeric(0),
  color = character(0),
  fill = character(0),
  linewidth = numeric(0),
  linetype = numeric(0),
  style = S7::class_missing,
  x = numeric(0),
  y = numeric(0),
  id = character(0),
  ...
)
}
\arguments{
\item{center}{\code{\link{ob_point}} at center of the rectangle}

\item{width}{width}

\item{height}{height}

\item{east}{right middle point (\code{\link{ob_point}})}

\item{north}{top middle point (\code{\link{ob_point}})}

\item{west}{left middle point (\code{\link{ob_point}})}

\item{south}{top middle point (\code{\link{ob_point}})}

\item{northeast}{upper right point (\code{\link{ob_point}})}

\item{northwest}{upper left point (\code{\link{ob_point}})}

\item{southwest}{lower left point (\code{\link{ob_point}})}

\item{southeast}{lower right point (\code{\link{ob_point}})}

\item{angle}{Rectangle rotation. \emph{Settable.}}

\item{vertex_radius}{A numeric or unit vector of length one, specifying the corner radius for rounded corners}

\item{label}{A character, angle, or \code{\link{ob_label}} object}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{a style object}

\item{x}{overrides x-coordinate in \code{center@x}}

\item{y}{overrides y-coordinate in \code{center@x}}

\item{id}{character string to identify object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to \code{style}}
}
\value{
\code{\link{ob_rectangle}} object
}
\description{
ob_rectangle class
}
\examples{
ggdiagram() +
  ob_rectangle(center = ob_point(0,0), width = 3, height = 2)
}
\section{Additional properties}{

\describe{
\item{\code{@xy}}{returns a matrix of xy coordinates of \code{center} points}

\item{\code{@aesthetics}}{A list of information about the object's aesthetic properties}

\item{\code{@area}}{returns rectangle area}

\item{\code{@bounding_box}}{returns the \code{ob_rectangle} that contains all the rectangles in the object}

\item{\code{@perimeter}}{returns the rectangle perimeter}

\item{\code{@side}}{returns the east, north, west, and south \code{ob_segment} of the rectangles}

\item{\code{@length}}{returns the number of rectangles in the object}

\item{\code{@geom}}{a function that returns a \code{ggforce::geom_shape} object}

\item{\code{@normal_at}}{A function that finds a point perpendicular to the rectangle at angle \code{theta} at the specified distance.}

\item{\code{@point_at}}{A function that finds a point on the rectangle at an angle \code{theta}}

\item{\code{@tangent_at}}{A function that finds a tangent line on the rectangle. Uses point_at to find the tangent point at angle theta and then returns the tangent line at that point. If a point is supplied instead of an angle, the point is projected onto the ellipse and then the tangent line is found from there.}

\item{\code{@tibble}}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggforce::geom_shape}}
}}

