% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{ob_reuleaux}
\alias{ob_reuleaux}
\title{Reuleaux polygon}
\usage{
ob_reuleaux(
  center = ob_point(0, 0),
  n = 5,
  radius = 1,
  angle = 90,
  label = character(0),
  vertex_radius = numeric(0),
  alpha = numeric(0),
  color = "black",
  fill = character(0),
  linewidth = numeric(0),
  linetype = numeric(0),
  style = S7::class_missing,
  id = character(0),
  ...
)
}
\arguments{
\item{center}{\code{\link{ob_point}} at center of the rectangle}

\item{n}{Number of sides. True Reuleaux polygons have an odd number of sides, but Reauleaux-like shapes with an even number of sides are possible.}

\item{radius}{Distance from center to a vertex}

\item{angle}{Angle of the object's rotation}

\item{label}{A character, angle, or \code{\link{ob_label}} object}

\item{vertex_radius}{A numeric or unit vector of length one, specifying the corner radius}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{Gets and sets the styles associated with polygons}

\item{id}{character string to identify object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> unused}
}
\value{
ob_reuleaux object
}
\description{
Reuleaux polygon
}
\examples{
ggdiagram() +
  ob_reuleaux(n = 3, fill = "royalblue", color = NA)
}
\section{Additional properties}{

\describe{
\item{\code{@aesthetics}}{A list of information about the object's aesthetic properties}

\item{\code{@angle_at}}{A function that finds the angle of the specified point in relation to the circle's center}

\item{\code{@arc_at}}{A function that finds the arc at a specified angle}

\item{\code{@arc_radius}}{The radius of the arcs used to construct the reuleaux object}

\item{\code{@arcs}}{Returns the arcs of each reuleaux object}

\item{\code{@bounding_box}}{a rectangle that contains all the reuleaux objects}

\item{\code{@central_angle}}{The angle from the center to adjacent vertices of the reuleaux object}

\item{\code{@chord_length}}{The length of each chord of the arcs used to construct the reuleaux object}

\item{\code{@circumference}}{circumference of the reuleaux object}

\item{\code{@circumscribed}}{Returns the circle that circumscribes the object}

\item{\code{@geom}}{A function that converts the object to a geom. Any additional parameters are passed to \code{ggforce::geom_shape}.}

\item{\code{@inscribed}}{Returns the circle that inscribes the object}

\item{\code{@inscribed_angle}}{The angle of the arcs used to construct the reuleaux object}

\item{\code{@length}}{The number of circles in the circle object}

\item{\code{@normal_at}}{A function that finds a point that is perpendicular from the circle and at a specified distance}

\item{\code{@point_at}}{A function that finds a point on the circle at the specified angle.}

\item{\code{@polygon}}{a tibble containing information to create all the polygon points in a reuleaux object}

\item{\code{@tangent_at}}{A function that finds the tangent line at the specified angle.}

\item{\code{@vertices}}{Returns the vertices of the reuleaux object}
}}

