% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{ob_shape_list}
\alias{ob_shape_list}
\title{ob_shape_list}
\usage{
ob_shape_list(.data = class_missing)
}
\arguments{
\item{.data}{a list of objects}
}
\description{
makes a heterogeneous list of different ggdiagram objects
}
