% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{ob_style}
\alias{ob_style}
\title{ob_style class}
\usage{
ob_style(
  alpha = class_missing,
  angle = class_missing,
  arrow_head = class_missing,
  arrow_fins = class_missing,
  arrow_mid = class_missing,
  color = class_missing,
  family = class_missing,
  fill = class_missing,
  fontface = class_missing,
  hjust = class_missing,
  justify = class_missing,
  label.color = class_missing,
  label.margin = class_missing,
  label.padding = class_missing,
  label.r = class_missing,
  label.size = class_missing,
  arrowhead_length = class_missing,
  length_head = class_missing,
  length_fins = class_missing,
  length_mid = class_missing,
  lineend = class_missing,
  lineheight = class_missing,
  linejoin = class_missing,
  linewidth_fins = class_missing,
  linewidth_head = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  n = class_missing,
  nudge_x = class_missing,
  nudge_y = class_missing,
  polar_just = class_missing,
  resect = class_missing,
  resect_fins = class_missing,
  resect_head = class_missing,
  shape = class_missing,
  size = class_missing,
  size.unit = class_missing,
  straight = class_missing,
  stroke = class_missing,
  stroke_color = class_missing,
  stroke_width = class_missing,
  text.color = class_missing,
  vjust = class_missing,
  ...
)
}
\arguments{
\item{alpha}{numeric value for alpha transparency}

\item{angle}{angle of text}

\item{arrow_head}{A 2-column matrix of polygon points}

\item{arrow_fins}{A 2-column matrix of polygon points}

\item{arrow_mid}{A 2-column matrix of polygon points}

\item{color}{character string for color}

\item{family}{font family}

\item{fill}{character string for fill color}

\item{fontface}{Can be plain, bold, italic, or bold.italic}

\item{hjust}{horizontal justification.}

\item{justify}{A numeric(1) between 0 and 1 to control where the arrows should be drawn relative to the path's endpoints. A value of 0 sets the arrow's tips at the path's end, whereas a value of 1 sets the arrow's base at the path's end. From ggarrow.}

\item{label.color}{Color of label outline.}

\item{label.margin}{Amount of distance around label. Unit vector of length four. Usually created with \code{ggplot2::margin}.}

\item{label.padding}{Amount of padding around label. Unit vector of length four. Usually created with \code{ggplot2::margin}.}

\item{label.r}{Radius of rounded corners. Defaults to 0.15 lines.}

\item{label.size}{Width of label outline.}

\item{arrowhead_length}{Determines the size of the arrow ornaments. This parameter becomes the \code{length} parameter in ggarrow functions. Numeric values set the ornament size relative to the linewidth. A unit value sets the ornament size in an absolute manner.}

\item{length_head}{Determines the size of the arrow head. Numeric values set the ornament size relative to the linewidth. A unit value sets the ornament size in an absolute manner. From ggarrow.}

\item{length_fins}{Determines the size of the arrow fins. Numeric values set the ornament size relative to the linewidth. A unit value sets the ornament size in an absolute manner. From ggarrow.}

\item{length_mid}{Determines the size of the middle arrows. Numeric values set the ornament size relative to the linewidth. A unit value sets the ornament size in an absolute manner. From ggarrow.}

\item{lineend}{Line end style (round, butt, square).}

\item{lineheight}{Height of line of text}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linewidth_fins}{Line width for arrow fins}

\item{linewidth_head}{Line width for arrow fins}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{n}{Number of points in a polygon, circle, arc, or ellipse}

\item{nudge_x}{Horizontal adjustment to nudge labels by.}

\item{nudge_y}{Vertical adjustment to nudge labels by.}

\item{polar_just}{an angle, polar point, or point that alters hjust and vjust (polar polar_just not stored in style)}

\item{resect}{A numeric(1) denoting millimeters or \if{html}{\out{<unit>}} to shorten the arrow head and fins.}

\item{resect_fins}{A numeric(1) denoting millimeters or \if{html}{\out{<unit>}} to shorten the arrow fins}

\item{resect_head}{A numeric(1) denoting millimeters or \if{html}{\out{<unit>}} to shorten the arrow head.}

\item{shape}{Point shape type. Can be specified with an integer (between 0 and 25), a single character (which uses that character as the plotting symbol), a . to draw the smallest rectangle that is visible (i.e., about one pixel), an NA to draw nothing, or a mapping to a discrete variable.}

\item{size}{numeric size}

\item{size.unit}{How the size aesthetic is interpreted: as points ("pt"), millimeters ("mm"), centimeters ("cm"), inches ("in"), or picas ("pc").}

\item{straight}{logical. If TRUE, make bzpath label text straight instead of curved.}

\item{stroke}{Width of point border line}

\item{stroke_color}{Color of point border line}

\item{stroke_width}{Stroke width in arrows}

\item{text.color}{Color of label text.}

\item{vjust}{vertical justification}

\item{...}{unused}
}
\description{
ob_style class
}
