% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{path}
\alias{path}
\title{Arrow path from one shape to another}
\usage{
path(x, y, ...)
}
\arguments{
\item{x}{first shape (e.g., point, circle, ellipse, rectangle)}

\item{y}{second shape}
}
\description{
Arrow path from one shape to another
}
