% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{place}
\alias{place}
\title{Place an object a specified distance from another object}
\usage{
place(x, from, where = "right", sep = 1, ...)
}
\arguments{
\item{x}{first shape (e.g., point, circle, ellipse, rectangle)}

\item{y}{second shape}
}
\description{
Place an object a specified distance from another object
}
