% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{polar2just}
\alias{polar2just}
\title{polar2just}
\usage{
polar2just(x, multiplier = 1.2, axis = c("h", "v"))
}
\arguments{
\item{x}{angle}

\item{multiplier}{distance}

\item{axis}{vertical (v) or horizontal (h)}
}
\description{
Convert hjust and vjust parameters from polar coordinates
}
