% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{projection}
\alias{projection}
\title{Find projection of a point on an object (e.g., line or segment)}
\usage{
projection(point, object, ...)
}
\arguments{
\item{point}{point}

\item{object}{object (e.g., line or segment)}
}
\description{
Find projection of a point on an object (e.g., line or segment)
}
