% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{rectangle}
\alias{rectangle}
\title{rectangle class}
\usage{
rectangle(
  center = class_missing,
  width = class_missing,
  height = class_missing,
  northeast = class_missing,
  northwest = class_missing,
  southwest = class_missing,
  southeast = class_missing,
  label = class_missing,
  alpha = class_missing,
  color = "black",
  fill = NA_character_,
  linewidth = 0.5,
  linetype = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{center}{point at center of the circle}

\item{width}{width}

\item{height}{height}

\item{style}{a style object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to style object if style is empty}

\item{xy}{2-column matrix of points}
}
\description{
rectangle class
}
\examples{
# specify center point and radius
p <- point(0,0)
rectangle(p, width = 2, height = 2)
}
