% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{redefault}
\alias{redefault}
\title{Make a variant of a function with alternate defaults}
\usage{
redefault(.f, ...)
}
\arguments{
\item{.f}{function}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> new defaults}
}
\value{
function
}
\description{
Makes a copy of a function with new defaults. Similar to \code{\link[purrr:partial]{purrr::partial}} except that arguments with new defaults still accept input.
}
\examples{
squircle <- redefault(ob_ellipse, m1 = 4)
squircle(a = 3)
}
