% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale an object in 2 dimensions}
\usage{
rescale(x, scale = 1, origin = ob_point(0, 0))
}
\arguments{
\item{x}{object}

\item{scale}{numeric value by which the object is rescaled}

\item{origin}{length 2 vector or point from which scaling occurs}
}
\value{
shape object
}
\description{
Rescale an object in 2 dimensions
}
