% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{resect}
\alias{resect}
\title{resect}
\usage{
resect(x, distance, ...)
}
\arguments{
\item{x}{object}

\item{distance}{resect distance}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}

\item{resect}{a numeric distance}
}
\value{
object of same class as x
}
\description{
Shorten segments
}
