% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate an object in 2 dimensions}
\usage{
rotate(x, theta, ..., origin = ob_point(0, 0))
}
\arguments{
\item{x}{object}

\item{theta}{angle}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}

\item{origin}{length 2 vector  or point about which rotation occurs}
}
\value{
shape object
}
\description{
Rotate an object in 2 dimensions
}
