% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{round_probability}
\alias{round_probability}
\title{Probability rounding}
\usage{
round_probability(
  p,
  accuracy = 0.01,
  digits = NULL,
  max_digits = NULL,
  remove_leading_zero = TRUE,
  round_zero_one = TRUE,
  phantom_text = NULL,
  phantom_color = NULL
)
}
\arguments{
\item{p}{probability}

\item{accuracy}{smallest increment}

\item{digits}{significant digits}

\item{max_digits}{maximum rounding digits}

\item{remove_leading_zero}{remove leading zero}

\item{round_zero_one}{round 0 and 1}

\item{phantom_text}{invisible text inserted on the right}

\item{phantom_color}{color of phantom text}
}
\value{
a character vector
}
\description{
Rounds to significant digits, removing leading zeros.
}
\examples{
round_probability(c(0, .0012, .012, .12, .99, .992, .9997, 1), digits = 2)
}
