% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segment}
\alias{segment}
\title{segment class}
\usage{
segment(
  p1 = class_missing,
  p2 = class_missing,
  alpha = class_missing,
  arrow_head = ggarrow::arrow_head_minimal(90),
  arrow_fins = class_missing,
  arrowhead_length = 4,
  length_head = class_missing,
  length_fins = class_missing,
  color = class_missing,
  lineend = class_missing,
  linejoin = class_missing,
  linewidth = 0.5,
  linewidth_fins = class_missing,
  linewidth_head = class_missing,
  linetype = class_missing,
  resect = class_missing,
  resect_fins = class_missing,
  resect_head = class_missing,
  stroke_color = class_missing,
  stroke_width = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{style}{a style list}

\item{...}{properties passed to style}

\item{b}{coefficient in general form: a * x + b * y + c = 0}

\item{a}{constant in general form: a * x + b * y + c = 0}

\item{slope}{coefficient in y = slope * x + intercept}

\item{intercept}{value of y when x is 0}

\item{xintercept}{value of x when y is 0}
}
\description{
segment class
}
