% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{signs_centered}
\alias{signs_centered}
\title{Centering signed numbers}
\usage{
signs_centered(x, space = NULL, encoding = "UTF-8", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{space}{a character to be added to negative numbers (defaults to a UTF-8 figure space)}

\item{encoding}{type of encoding (defaults to UTF-8)}

\item{...}{parameters passed to signs:signs}
}
\value{
a vector of numbers converted to characters
}
\description{
A wrapper function for the signs::signs function. It adds a space to the right side of negative numbers so that it appear as if the minus sign does not affect the number's centering.
}
