% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\docType{data}
\name{str}
\alias{str}
\title{structure}
\format{
An object of class \code{S7_external_generic} of length 4.
}
\usage{
str
}
\arguments{
\item{object}{object}
}
\description{
structure
}
\keyword{internal}
