% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcs.R
\name{arc}
\alias{arc}
\title{arc class}
\usage{
arc(
  center = point(0, 0),
  radius = 1,
  start = 0,
  end = 0,
  label = class_missing,
  start_point = class_missing,
  end_point = class_missing,
  n = 360,
  wedge = FALSE,
  alpha = class_missing,
  arrow_head = class_missing,
  arrow_fins = class_missing,
  arrowhead_length = class_missing,
  length_head = class_missing,
  length_fins = class_missing,
  color = class_missing,
  fill = class_missing,
  lineend = class_missing,
  linejoin = class_missing,
  linewidth = 0.25,
  linewidth_fins = class_missing,
  linewidth_head = class_missing,
  linetype = class_missing,
  resect = class_missing,
  resect_fins = class_missing,
  resect_head = class_missing,
  stroke_color = class_missing,
  stroke_width = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{center}{point at center of the arc (default = point(0,0))}

\item{radius}{distance between center and edge arc (default = 1)}

\item{start}{start angle (default = 0 degrees)}

\item{end}{end angle (default = 0 degrees)}

\item{label}{A character, angle, or label object}

\item{n}{number of points in arc (default = 360)}

\item{style}{a style object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to style object if style is empty}

\item{theta}{interior angle (end - start)}

\item{length}{The number of arcs in the arc object}
}
\description{
arc class
}
\examples{
# specify center point and radius
p <- point(0,0)
arc(p, radius = 6, start = degree(0), end = degree(30))
}
