% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{bind}
\alias{bind}
\title{bind method}
\usage{
bind(x, ...)
}
\arguments{
\item{x}{list of objects to bind}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\value{
a bound object of same class as x (or list of objects if x contains objects of different types)
}
\description{
bind method
}
\examples{
bind(c(ob_point(1,2), ob_point(3,4)))
bind(c(ob_circle(ob_point(0,0), radius = 1),
       ob_circle(ob_point(1,1), radius = 2)))
}
