% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{circle_from_3_points}
\alias{circle_from_3_points}
\title{Get a circle from 3 points}
\usage{
circle_from_3_points(p1, p2 = NULL, p3 = NULL, ...)
}
\arguments{
\item{p1}{an ob_point of length 1 or length 3}

\item{p2}{an ob_point of length 1 or NULL}

\item{p3}{an ob_point of length 1 or NULL}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Pass arguments to ob_circle}
}
\value{
ob_point object
}
\description{
Get a circle from 3 points
}
\examples{
circle_from_3_points(ob_point(1,1),
                     ob_point(2,4),
                     ob_point(5,3))
}
