% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{connect}
\alias{connect}
\title{Arrow connect one shape to another}
\usage{
connect(from, to, ...)
}
\arguments{
\item{from}{first shape object}

\item{to}{second shape object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Arguments passed to style}
}
\value{
ob_segment
}
\description{
Arrow connect one shape to another
}
