% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{data2shape}
\alias{data2shape}
\title{Make shapes from data}
\usage{
data2shape(data, shape)
}
\arguments{
\item{data}{data.frame or tibble}

\item{shape}{shape function}
}
\value{
shape object
}
\description{
Allows a data.frame or tibble to be converted to shape objects.
}
\examples{
d <- data.frame(
  x = 1:2,
  y = 1:2,
  fill = c("blue", "forestgreen"),
  color = NA,
  radius = c(.25,0.5))

ggdiagram() +
  data2shape(d, ob_circle)
}
