% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{ellipse}
\alias{ellipse}
\title{ellipse class}
\usage{
ellipse(
  center = point(0, 0),
  a = 1,
  b = 1,
  angle = 0,
  m1 = 2,
  m2 = 2,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  n = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{center}{point at center of ellipse}

\item{a}{distance of semi-major axis}

\item{b}{distance of semi-minor axis}

\item{m1}{exponent of semi-major axis}

\item{m2}{exponent of semi-minor axis}

\item{n}{number of points in ellipse (default = 360)}

\item{style}{a style object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to style object if style is empty}

\item{theta}{ellipse rotation}
}
\description{
ellipse class
}
\examples{
# specify center point and semi-major axes
p <- point(0,0)
ellipse(p, a = 2, b = 3)
}
