% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{equation}
\alias{equation}
\title{equation}
\usage{
equation(
  x,
  type = c("y", "general", "parametric"),
  output = c("markdown", "latex"),
  digits = 2
)
}
\arguments{
\item{x}{object}

\item{type}{equation type. Can be \code{y} (default), \code{general}, or \code{parametric}}

\item{output}{Can be \code{markdown} (default) or \code{latex}}

\item{digits}{rounding digits}
}
\value{
string
}
\description{
Get equation for object
}
