% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{equation}
\alias{equation}
\title{equation}
\usage{
equation(x, type = c("y", "general", "parametric"), digits = 2)
}
\arguments{
\item{x}{object}

\item{type}{equation type. Can be \code{y}, \code{general}, or \code{parametric}}

\item{digits}{rounding digits}
}
\description{
Get equation for object
}
