% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{equation}
\alias{equation}
\title{equation}
\usage{
equation(
  x,
  type = c("y", "general", "parametric"),
  output = c("markdown", "latex"),
  digits = 2
)
}
\arguments{
\item{x}{object}

\item{type}{equation type. Can be \code{y} (default), \code{general}, or \code{parametric}}

\item{output}{Can be \code{markdown} (default) or \code{latex}}

\item{digits}{rounding digits}
}
\value{
string
}
\description{
Get equation for object
}
\examples{
l1 <- ob_line(slope = 2, intercept = 4)
c1 <- ob_circle(radius = 3)
ggdiagram() +
  l1 +
  c1 +
  ob_label(label = equation(c1),
           center = c1@center,
           size = 16) +
  ob_label(label = equation(l1),
           center = ob_segment(intersection(l1, c1))@midpoint(),
           angle = l1@angle,
           size = 16) +
 ggplot2::theme_minimal(base_size = 20)
}
