% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram-package.R
\docType{package}
\name{ggdiagram-package}
\alias{ggdiagram-package}
\title{ggdiagram: Object-Oriented Diagram Plots with 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates diagrams with an object-oriented approach. Geometric objects have computed properties with information about themselves (e.g., their area) or about their relationships with other objects (e.g, the distance between their edges). The objects have methods to convert them to geoms that can be plotted in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjschne/ggdiagram}
  \item \url{https://wjschne.github.io/ggdiagram/}
  \item Report bugs at \url{https://github.com/wjschne/ggdiagram/issues}
}

}
\author{
\strong{Maintainer}: W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})

}
\keyword{internal}
