% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{ggdiagram}
\alias{ggdiagram}
\title{Create a new ggdiagram}
\usage{
ggdiagram(xlim = c(-50, 50), ylim = c(-50, 50))
}
\arguments{
\item{xlim}{Numerical vector of length 2 specifying xmin and xmax}

\item{ylim}{Numerical vector of length 2 specifying ymin and ymax}
}
\description{
Create a new ggdiagram
}
