% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{ggdiagram}
\alias{ggdiagram}
\title{ggdiagram function}
\usage{
ggdiagram(
  base_size = 11,
  base_family = "sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  theme_function = ggplot2::theme_void,
  ...
)
}
\arguments{
\item{family}{font family}
}
\description{
This is a convenient way to specify
}
\examples{
ggdiagram() + ob_circle()
}
