% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{ggdiagram}
\alias{ggdiagram}
\title{ggdiagram function}
\usage{
ggdiagram(
  font_family = "sans",
  font_size = 11,
  linewidth = 0.5,
  point_size = 1.5,
  rect_linewidth = linewidth,
  theme_function = ggplot2::theme_void,
  ...
)
}
\arguments{
\item{font_family}{font family}

\item{font_size}{font size in points}

\item{linewidth}{line width}

\item{point_size}{point size}

\item{rect_linewidth}{line width of rectangles}

\item{theme_function}{A complete \link[ggplot2:ggtheme]{ggplot2 theme} function (e.g., \link[ggplot2:ggtheme]{ggplot2::theme_minimal}). Defaults to \link[ggplot2:ggtheme]{ggplot2::theme_void}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Arguments sent to \link[ggplot2:theme]{ggplot2::theme}}
}
\value{
ggplot function
}
\description{
This is a convenient way to specify geom defaults
}
\examples{
ggdiagram(font_size = 20, font_family = "serif", linewidth = 3) +
   ob_circle(label = "Circle") +
   ob_rectangle(label = "Rectangle", x = 3, width = 3)
}
