% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersection_angle}
\alias{intersection_angle}
\title{Compute the angle of the intersection of two objects}
\usage{
intersection_angle(x, y)
}
\arguments{
\item{x}{an object (e.g., \link{ob_point}, \link{ob_segment}, \link{ob_line})}

\item{y}{an object (e.g., \link{ob_point}, \link{ob_segment}, \link{ob_line})}
}
\value{
\link{ob_angle} object
}
\description{
Compute the angle of the intersection of two objects
}
