% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{map_ob}
\alias{map_ob}
\title{map_ob}
\usage{
map_ob(.x, .f, ..., .progress = FALSE)
}
\arguments{
\item{.x}{a ggdiagram object}

\item{.f}{a function that returns a ggdiagram object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to .f}

\item{.progress}{display progress if TRUE}
}
\value{
a ggdiagram object
}
\description{
A wrapper for \link[purrr:map]{purrr::map}. It takes a ggdiagram object with multiple elements, applies a function to each element within the object, and returns a ggdiagram object
}
