% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{nudge}
\alias{nudge}
\title{Move an object}
\usage{
nudge(object, x, y, ...)
}
\arguments{
\item{object}{object}

\item{x}{nudge right and left}

\item{y}{nudge up and down}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\value{
object of same class as \code{object}
}
\description{
Move an object
}
\examples{
ob_circle() |> nudge(x = 2)
# Alternative to nudge:
ob_circle() + ob_point(2, 0)
}
