% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{ob_intercept}
\alias{ob_intercept}
\title{ob_intercept}
\usage{
ob_intercept(
  center = ob_point(0, 0),
  width = 1,
  label = character(0),
  top = S7::class_missing,
  left = S7::class_missing,
  right = S7::class_missing,
  vertex_radius = numeric(0),
  alpha = numeric(0),
  color = character(0),
  fill = character(0),
  linewidth = numeric(0),
  linetype = numeric(0),
  x = numeric(0),
  y = numeric(0),
  style = S7::class_missing,
  id = character(0),
  ...
)
}
\arguments{
\item{center}{\code{\link{ob_point}} at center}

\item{width}{length of side}

\item{label}{A character, angle, or \code{\link{ob_label}} object}

\item{top}{Top vertex of triangle}

\item{left}{Left vertex of triangle}

\item{right}{Right vertex of triangle}

\item{vertex_radius}{A numeric or unit vector of length one, specifying the vertex radius}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{x}{overrides x-coordinate in \code{center@x}}

\item{y}{overrides x-coordinate in \code{center@y}}

\item{style}{Gets and sets the styles associated with polygons}

\item{id}{character string to identify object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\value{
ob_polygon object
}
\description{
Triangle polygons used in path diagrams.
}
\section{Slots}{

\describe{
\item{\code{length}}{The number of polygons in the ob_polygon object}

\item{\code{tibble}}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggplot2::geom_polygon}.}
}}

