% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{ob_latex}
\alias{ob_latex}
\title{ob_latex class}
\usage{
ob_latex(
  tex = character(0),
  center = ob_point(0, 0),
  width = numeric(0),
  height = numeric(0),
  hjust = 0.5,
  vjust = 0.5,
  angle = 0,
  aspect_ratio = 1,
  border = numeric(0),
  family = character(0),
  math_mode = TRUE,
  filename = character(0),
  color = character(0),
  fill = "white",
  density = 300,
  latex_packages = character(0),
  preamble = character(0),
  force_recompile = TRUE,
  delete_files = TRUE,
  id = character(0)
)
}
\arguments{
\item{tex}{LaTeX equation}

\item{center}{An \link{ob_point}}

\item{width}{width (specify width or height but not both)}

\item{height}{height (specify width or height but not both)}

\item{hjust}{horizontal adjustment. 0 means left justified, 1 means right justified, 0.5 means centered}

\item{vjust}{vertical justification. 0 means bottom aligned, 1 means top aligned, 0.5 means vertically centered}

\item{angle}{angle of text}

\item{aspect_ratio}{alters the aspect ratio of the image}

\item{border}{border space (in points) around image}

\item{family}{font family (installed on system) of plain text}

\item{math_mode}{include dollar signs automatically. Set to \code{FALSE} when the latex command is not in math mode}

\item{filename}{bare file name without extension (e.g., \code{myequation})}

\item{color}{set color of equation text}

\item{fill}{set color of background rectangle}

\item{density}{image quality (dots per inch)}

\item{latex_packages}{load latex packages}

\item{preamble}{additional latex commands to load in preamble}

\item{force_recompile}{Will re-run xelatex even if .pdf file exists already}

\item{delete_files}{Delete .tex and .pdf files after image is generated.}

\item{id}{character string to identify object}
}
\value{
ob_latex object
}
\description{
make a latex equation
}
\section{Slots}{

\describe{
\item{\code{rectangle}}{gets or sets rectangle that contains the image}

\item{\code{image}}{raster bitmap}
}}

