% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{ob_polygon}
\alias{ob_polygon}
\title{The ob_polygon class}
\usage{
ob_polygon(
  p = class_missing,
  label = character(0),
  vertex_radius = numeric(0),
  alpha = numeric(0),
  color = character(0),
  fill = character(0),
  linewidth = numeric(0),
  linetype = numeric(0),
  style = class_missing,
  ...
)
}
\arguments{
\item{p}{ob_point or list of ob_point objects}

\item{label}{A character, angle, or label object}

\item{vertex_radius}{A numeric or unit vector of length one, specifying the corner radius}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{Gets and sets the styles associated with polygons}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\value{
ob_polygon object
}
\description{
A polygon is specified with an ob_point that contains at least 3 points, the start and the end. Any number of intermediate points are possible.
}
\details{
If you wish to specify multiple polygons, you must supply a list of ob_points. When plotted, the ob_polygon function uses the ggforce::geom_shape function to create the geom.
}
\section{Slots}{

\describe{
\item{\code{length}}{The number of polygons in the ob_polygon object}

\item{\code{tibble}}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggforce::geom_shape}.}
}}

