% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{ob_rectangle}
\alias{ob_rectangle}
\title{ob_rectangle class}
\usage{
ob_rectangle(
  center = class_missing,
  width = class_missing,
  height = class_missing,
  east = class_missing,
  north = class_missing,
  west = class_missing,
  south = class_missing,
  northeast = class_missing,
  northwest = class_missing,
  southwest = class_missing,
  southeast = class_missing,
  angle = 0,
  vertex_radius = class_missing,
  label = class_missing,
  alpha = class_missing,
  color = "black",
  fill = NA_character_,
  linewidth = class_missing,
  linetype = class_missing,
  style = class_missing,
  x0 = class_missing,
  y0 = class_missing,
  ...
)
}
\arguments{
\item{center}{point at center of the circle}

\item{width}{width}

\item{height}{height}

\item{east}{right middle point}

\item{north}{top middle point}

\item{west}{left middle point}

\item{south}{top middle point}

\item{northeast}{upper right point}

\item{northwest}{upper left point}

\item{southwest}{lower left point}

\item{southeast}{lower right point}

\item{angle}{angle of text}

\item{vertex_radius}{A numeric or unit vector of length one, specifying the corner radius for rounded corners}

\item{label}{A character, angle, or label object}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{a style object}

\item{x0}{overrides x-coordinate in \code{center@x}}

\item{y0}{overrides y-coordinate in \code{center@x}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to style object}
}
\description{
ob_rectangle class
}
\examples{
# specify center point
p <- ob_point(0,0)
ob_rectangle(p, width = 2, height = 2)
}
