% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{ob_reuleaux}
\alias{ob_reuleaux}
\title{Reuleaux polygon}
\usage{
ob_reuleaux(
  center = ob_point(0, 0),
  n = 5,
  radius = 1,
  angle = 90,
  label = character(0),
  vertex_radius = numeric(0),
  alpha = numeric(0),
  color = "black",
  fill = character(0),
  linewidth = numeric(0),
  linetype = numeric(0),
  style = S7::class_missing,
  id = character(0),
  ...
)
}
\arguments{
\item{center}{\code{\link{ob_point}} at center of the rectangle}

\item{n}{Number of sides. True Reuleaux polygons have an odd number of sides, but Reauleaux-like shapes with an even number of sides are possible.}

\item{radius}{Distance from center to a vertex}

\item{angle}{angle of text}

\item{label}{A character, angle, or \code{\link{ob_label}} object}

\item{vertex_radius}{A numeric or unit vector of length one, specifying the corner radius}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{Gets and sets the styles associated with polygons}

\item{id}{character string to identify object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> unused}
}
\value{
ob_reuleaux object
}
\description{
Reuleaux polygon
}
