% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{ob_shape_list}
\alias{ob_shape_list}
\title{ob_shape_list}
\usage{
ob_shape_list(.data = list())
}
\arguments{
\item{.data}{a list of objects}
}
\value{
An object of \code{\link{ob_shape_list}} class. List of objects that can be converted to geoms
}
\description{
makes a heterogeneous list of different ggdiagram objects
}
