% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path}
\alias{path}
\title{The path class}
\usage{
path(
  p = class_missing,
  label = class_missing,
  alpha = class_missing,
  arrow_head = class_missing,
  arrow_fins = class_missing,
  arrowhead_length = class_missing,
  length_head = class_missing,
  length_fins = class_missing,
  color = class_missing,
  fill = class_missing,
  lineend = class_missing,
  linejoin = class_missing,
  linewidth = 0.75,
  linewidth_fins = class_missing,
  linewidth_head = class_missing,
  linetype = class_missing,
  resect = class_missing,
  resect_fins = class_missing,
  resect_head = class_missing,
  stroke_color = class_missing,
  stroke_width = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{p}{point object or list of point objects}

\item{style}{Gets and sets the styles associated with paths}

\item{...}{properties passed to style}

\item{length}{The number of paths in the path object}

\item{tibble}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggarrow::geom_arrow}.}
}
\description{
A path is specified with a point object that contains at least 2 points, the start and the end. Any number of intermediate points are possible.
}
\details{
If you wish to specify multiple paths, you must supply a list of point objects. When plotted, the path function uses the ggarrow::geom_arrow function to create the geom.
}
