% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{pgon}
\alias{pgon}
\title{The pgon (polygon) class}
\usage{
pgon(
  p = class_missing,
  label = class_missing,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  linewidth = 0.75,
  linetype = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{p}{point object or list of point objects}

\item{label}{A character, angle, or label object}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{Gets and sets the styles associated with polygons}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}
}
\description{
A polygon is specified with a point object that contains at least 3 points, the start and the end. Any number of intermediate points are possible. The pgon class could not be named 'polygon' because that function name is used in base R graphics.
}
\details{
If you wish to specify multiple polygons, you must supply a list of point objects. When plotted, the pgon function uses the ggplot2::geom_polygon function to create the geom.
}
\section{Slots}{

\describe{
\item{\code{length}}{The number of polygons in the pgon object}

\item{\code{tibble}}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggplot2::geom_polygon}.}
}}

