% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{place}
\alias{place}
\title{Place an object a specified distance from another object}
\usage{
place(x, from, where = "right", sep = 1, ...)
}
\arguments{
\item{x}{shape object}

\item{from}{shape that x is placed in relation to}

\item{where}{named direction, angle, or number (degrees)}

\item{sep}{separation distance}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Arguments passed to \link{ob_style}}
}
\value{
object of same class as \code{x}
}
\description{
Place an object a specified distance from another object
}
