% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{point}
\alias{point}
\alias{polar}
\title{point}
\usage{
point(
  x = 0,
  y = 0,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  shape = class_missing,
  size = class_missing,
  stroke = class_missing,
  style = class_missing,
  ...
)

polar(
  theta = class_missing,
  r = class_missing,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  shape = class_missing,
  size = class_missing,
  stroke = class_missing,
  style = class_missing
)
}
\arguments{
\item{x}{Vector of coordinates on the x-axis (also can take a tibble/data.frame or 2-column matrix as input.)}

\item{y}{Vector of coordinates on the y-axis}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{shape}{Point shape type. Can be specified with an integer (between 0 and 25), a single character (which uses that character as the plotting symbol), a . to draw the smallest rectangle that is visible (i.e., about one pixel), an NA to draw nothing, or a mapping to a discrete variable.}

\item{size}{numeric size}

\item{stroke}{Width of point border line}

\item{style}{Gets and sets the styles associated with points}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style}

\item{theta}{Angle of the vector from the origin to the point}

\item{r}{Radius = Distance from the origin to the point}
}
\description{
Points are specified with x and y coordinates.

Polar points are ordinary points but are specified with an angle (theta) and a radial distance (r)
}
\section{Slots}{

\describe{
\item{\code{auto_label}}{Gets x and y coordinates and makes a label \code{"(x,y)"}}

\item{\code{length}}{The number of points in the point object}

\item{\code{tibble}}{Gets a tibble (data.frame) containing parameters and styles used by \code{ggplot2::geom_point}.}

\item{\code{xy}}{Gets a 2-column matrix of the x and y coordinates of the point object.}

\item{\code{geom}}{A function that converts the object to a geom. Any additional parameters are passed to \code{ggplot2::geom_point}.}
}}

