% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{point}
\alias{point}
\alias{polar}
\title{point}
\usage{
point(
  x = class_missing,
  y = class_missing,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  shape = class_missing,
  size = class_missing,
  stroke = class_missing,
  style = class_missing,
  ...
)

polar(theta = class_missing, r = class_missing, style = class_missing, ...)
}
\arguments{
\item{x}{Coordinate on the x-axis}

\item{y}{Coordinate on the y-axis}

\item{style}{a style object}

\item{...}{properties passed to style}

\item{theta}{Angle of the vector from the origin to the point}

\item{r}{Radius = Distance from the origin to the point}
}
\description{
Points are specified with x and y coordinates.# Polar class ----

Polar points are ordinary points but are specified with an angle (theta) and a radial distance (r)
}
