% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{projection}
\alias{projection}
\title{Find projection of a point on an object (e.g., line or segment)}
\usage{
projection(p, object, ...)
}
\arguments{
\item{p}{ob_point}

\item{object}{object (e.g., line or segment)}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> properties passed to style object}
}
\value{
ob_point
}
\description{
Find projection of a point on an object (e.g., line or segment)
}
