% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{rectangle}
\alias{rectangle}
\title{rectangle class}
\usage{
rectangle(
  center = class_missing,
  width = class_missing,
  height = class_missing,
  northeast = class_missing,
  northwest = class_missing,
  southwest = class_missing,
  southeast = class_missing,
  angle = 0,
  radius = class_missing,
  label = class_missing,
  alpha = class_missing,
  color = "black",
  fill = NA_character_,
  linewidth = 0.5,
  linetype = class_missing,
  style = class_missing,
  x = class_missing,
  y = class_missing,
  ...
)
}
\arguments{
\item{center}{point at center of the circle}

\item{width}{width}

\item{height}{height}

\item{northeast}{upper right point}

\item{northwest}{upper left point}

\item{southwest}{lower left point}

\item{southeast}{lower right point}

\item{angle}{angle of text}

\item{label}{A character, angle, or label object}

\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{fill}{character string for fill color}

\item{linewidth}{Width of lines}

\item{linetype}{type of lines}

\item{style}{a style object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to style object}
}
\description{
rectangle class
}
\examples{
# specify center point and radius
p <- point(0,0)
rectangle(p, width = 2, height = 2)
}
