% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{prop}
\alias{set_props}
\alias{props}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{S7}{\code{\link[S7]{prop}}, \code{\link[S7]{props}}, \code{\link[S7:props]{set_props}}}
}}

