% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_early.R
\name{unbind}
\alias{unbind}
\title{unbind}
\usage{
unbind(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> additional arguments (not used at this time)}
}
\value{
a list of objects, each of length 1
}
\description{
Converts an object with k elements into a list of k objects
}
